Adding Aws Security Credential Providers
